data_collection/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.1.0
numpy>=1.24.0
lxml>=4.9.0
python-dateutil>=2.8.0
pyyaml>=6.0
fake-useragent>=1.4.0
webdriver-manager>=4.0.0
